# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neelgund Developers API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads (all for admin/manager, own for employee)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadRequest"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /leads/{id}/meetings:
    get:
      operationId: listLeadMeetings
      tags: [lead-meetings]
      summary: List meetings for a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadMeeting"
    post:
      operationId: createLeadMeeting
      tags: [lead-meetings]
      summary: Schedule a meeting for a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadMeetingRequest"
      responses:
        "201":
          description: Meeting scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadMeeting"
  /leads/{id}/meetings/{meetingId}:
    put:
      operationId: updateLeadMeeting
      tags: [lead-meetings]
      summary: Update a meeting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: meetingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadMeetingRequest"
      responses:
        "200":
          description: Meeting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadMeeting"
    delete:
      operationId: deleteLeadMeeting
      tags: [lead-meetings]
      summary: Delete a meeting
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: meetingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /leads/{id}/documents:
    get:
      operationId: listLeadDocuments
      tags: [lead-documents]
      summary: List documents for a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lead documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: addLeadDocument
      tags: [lead-documents]
      summary: Add a document to a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDocumentRequest"
      responses:
        "201":
          description: Document added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
          format: uuid
        employeeName:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        propertyInterest:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, follow_up, meeting_scheduled, negotiation, closed_won, closed_lost]
        notes:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - employeeId
        - name
        - phone
        - status
        - createdAt
        - updatedAt
    CreateLeadRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        propertyInterest:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, follow_up, meeting_scheduled, negotiation, closed_won, closed_lost]
        notes:
          type: string
          nullable: true
        latitude:
          type: number
          description: GPS latitude — required, captured at moment of lead entry
        longitude:
          type: number
          description: GPS longitude — required, captured at moment of lead entry
      required:
        - name
        - phone
        - latitude
        - longitude
    UpdateLeadRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        propertyInterest:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, follow_up, meeting_scheduled, negotiation, closed_won, closed_lost]
        notes:
          type: string
          nullable: true
    LeadMeeting:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        scheduledAt:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - leadId
        - scheduledAt
        - createdAt
    CreateLeadMeetingRequest:
      type: object
      properties:
        scheduledAt:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - scheduledAt
    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        mimeType:
          type: string
          nullable: true
        uploadedBy:
          type: string
          format: uuid
        createdAt:
          type: string
      required:
        - id
        - name
        - url
        - uploadedBy
        - createdAt
    AddDocumentRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        mimeType:
          type: string
          nullable: true
      required:
        - name
        - url
...